remove redundant icons from sidebar
